+++ blueprint/billing/routes/billingPortal.routes.ts
@@ -4,12 +4,12 @@ import { Metrics } from '@forklaunch/blueprint-monitoring';
 import { OpenTelemetryCollector } from '@forklaunch/core/http';
 import { ScopedDependencyFactory } from '@forklaunch/core/services';
 import { BillingPortalController } from '../controllers/billingPortal.controller';
-import { ServiceDependencies } from '../registrations';
+import { SchemaDependencies } from '../registrations';
 
 export const BillingPortalRoutes = (
   scopedServiceFactory: ScopedDependencyFactory<
     SchemaValidator,
-    ServiceDependencies,
+    SchemaDependencies,
     'BillingPortalService'
   >,
   openTelemetryCollector: OpenTelemetryCollector<Metrics>