+++ blueprint/sample-worker/routes/sampleWorker.routes.ts
@@ -6,14 +6,14 @@ import {
 import { forklaunchRouter, SchemaValidator } from '@forklaunch/blueprint-core';
 import { Metrics } from '@forklaunch/blueprint-monitoring';
 import { SampleWorkerController } from '../controllers/sampleWorker.controller';
-import { ServiceDependencies } from '../registrations';
+import { SchemaDependencies } from '../registrations';
 
 // returns an object with the router and the sampleWorkerGet and sampleWorkerPost methods for easy installation
 export const SampleWorkerRoutes = (
-  scopeFactory: () => ConfigInjector<SchemaValidator, ServiceDependencies>,
+  scopeFactory: () => ConfigInjector<SchemaValidator, SchemaDependencies>,
   scopedServiceFactory: ScopedDependencyFactory<
     SchemaValidator,
-    ServiceDependencies,
+    SchemaDependencies,
     'SampleWorkerService'
   >,
   openTelemetryCollector: OpenTelemetryCollector<Metrics>