+++ framework/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @forklaunch/core
 
+## 0.6.2
+
+### Patch Changes
+
+- Update package versions
+- Updated dependencies
+  - @forklaunch/validator@0.4.10
+  - @forklaunch/common@0.2.6
+
 ## 0.6.1
 
 ### Patch Changes